Rename to strobe
